update to origin

+++ jraft-rheakv/rheakv-core/src/test/java/com/alipay/sofa/jraft/rhea/storage/rhea/RheaKVTestCluster.java
@@ -45,22 +45,22 @@ public class RheaKVTestCluster {
 
     public static String          DB_PATH        = "rhea_db";
     public static String          RAFT_DATA_PATH = "rhea_raft";
-    public static Long[]          REGION_IDS     = new Long[] { 1L, 2L, 3L };
+    public static Long[]          REGION_IDS     = new Long[] { 1L, 2L };
 
-    private static final String[] CONF = {
-            "/conf/rhea_test_cluster_1.yaml",
-            "/conf/rhea_test_cluster_2.yaml",
-            "/conf/rhea_test_cluster_3.yaml"
-    };
+    private static final String[] CONF           = { "/conf/rhea_test_cluster_1.yaml", "/conf/rhea_test_cluster_2.yaml" };
 
     private List<RheaKVStore>     stores         = new CopyOnWriteArrayList<>();
 
     protected void start(final StorageType storageType) throws Exception {
-        deleteFiles();
+        start(storageType, true);
+    }
+
+    protected void start(final StorageType storageType, final boolean deleteFiles) throws Exception {
+        if (deleteFiles) {
+            deleteFiles();
+        }
         for (final String c : CONF) {
-            final ObjectMapper mapper = new ObjectMapper(new YAMLFactory());
-            final InputStream in = RheaKVTestCluster.class.getResourceAsStream(c);
-            final RheaKVStoreOptions opts = mapper.readValue(in, RheaKVStoreOptions.class);
+            final RheaKVStoreOptions opts = readOpts(c);
             opts.getStoreEngineOptions().setStorageType(storageType);
             final RheaKVStore rheaKVStore = new DefaultRheaKVStore();
             if (rheaKVStore.init(opts)) {
@@ -75,11 +75,25 @@ public class RheaKVTestCluster {
     }
 
     protected void shutdown() throws Exception {
+        shutdown(true);
+    }
+
+    protected void shutdown(final boolean deleteFiles) throws Exception {
         for (final RheaKVStore store : stores) {
             store.shutdown();
         }
-        deleteFiles();
+        stores.clear();
+        if (deleteFiles) {
+            deleteFiles();
+        }
         LOG.info("RheaKVTestCluster shutdown complete");
+    }
+
+    private RheaKVStoreOptions readOpts(final String conf) throws IOException {
+        final ObjectMapper mapper = new ObjectMapper(new YAMLFactory());
+        try (final InputStream in = RheaKVTestCluster.class.getResourceAsStream(conf)) {
+            return mapper.readValue(in, RheaKVStoreOptions.class);
+        }
     }
 
     protected RheaKVStore getRandomLeaderStore() {
@@ -95,7 +109,9 @@ public class RheaKVTestCluster {
             }
             try {
                 Thread.sleep(100);
-            } catch (InterruptedException ignored) {}
+            } catch (InterruptedException ignored) {
+                // ignored
+            }
         }
         throw new NotLeaderException("no leader on region: " + regionId);
     }
